fix: FoD: Refactor class names to represent commands (resolves #362)
