+++ fcli-core/fcli-fod/src/main/java/com/fortify/cli/fod/microservice/cli/mixin/FoDMicroserviceResolverMixin.java
@@ -1,13 +1,13 @@
 /*******************************************************************************
  * Copyright 2021, 2023 Open Text.
  *
- * The only warranties for products and services of Open Text 
- * and its affiliates and licensors ("Open Text") are as may 
- * be set forth in the express warranty statements accompanying 
- * such products and services. Nothing herein should be construed 
- * as constituting an additional warranty. Open Text shall not be 
- * liable for technical or editorial errors or omissions contained 
- * herein. The information contained herein is subject to change 
+ * The only warranties for products and services of Open Text
+ * and its affiliates and licensors ("Open Text") are as may
+ * be set forth in the express warranty statements accompanying
+ * such products and services. Nothing herein should be construed
+ * as constituting an additional warranty. Open Text shall not be
+ * liable for technical or editorial errors or omissions contained
+ * herein. The information contained herein is subject to change
  * without notice.
  *******************************************************************************/
 
@@ -23,7 +23,7 @@ import picocli.CommandLine.Mixin;
 import picocli.CommandLine.Option;
 import picocli.CommandLine.Parameters;
 
-public class FoDAppMicroserviceResolverMixin {
+public class FoDMicroserviceResolverMixin {
     public static abstract class AbstractFoDAppMicroserviceResolverMixin {
         @Mixin private FoDDelimiterMixin delimiterMixin;
         public abstract String getAppMicroserviceNameOrId();
