+++ fcli-core/fcli-fod/src/main/java/com/fortify/cli/fod/release/cli/cmd/FoDReleaseGetCommand.java
@@ -1,13 +1,13 @@
 /*******************************************************************************
  * Copyright 2021, 2023 Open Text.
  *
- * The only warranties for products and services of Open Text 
- * and its affiliates and licensors ("Open Text") are as may 
- * be set forth in the express warranty statements accompanying 
- * such products and services. Nothing herein should be construed 
- * as constituting an additional warranty. Open Text shall not be 
- * liable for technical or editorial errors or omissions contained 
- * herein. The information contained herein is subject to change 
+ * The only warranties for products and services of Open Text
+ * and its affiliates and licensors ("Open Text") are as may
+ * be set forth in the express warranty statements accompanying
+ * such products and services. Nothing herein should be construed
+ * as constituting an additional warranty. Open Text shall not be
+ * liable for technical or editorial errors or omissions contained
+ * herein. The information contained herein is subject to change
  * without notice.
  *******************************************************************************/
 
@@ -17,7 +17,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fortify.cli.common.output.cli.mixin.OutputHelperMixins;
 import com.fortify.cli.common.output.transform.IRecordTransformer;
 import com.fortify.cli.fod._common.output.cli.AbstractFoDJsonNodeOutputCommand;
-import com.fortify.cli.fod.release.cli.mixin.FoDAppRelResolverMixin;
+import com.fortify.cli.fod.release.cli.mixin.FoDReleaseResolverMixin;
 import com.fortify.cli.fod.release.helper.FoDAppRelHelper;
 
 import kong.unirest.UnirestInstance;
@@ -26,9 +26,9 @@ import picocli.CommandLine.Command;
 import picocli.CommandLine.Mixin;
 
 @Command(name = OutputHelperMixins.Get.CMD_NAME)
-public class FoDAppRelGetCommand extends AbstractFoDJsonNodeOutputCommand implements IRecordTransformer {
+public class FoDReleaseGetCommand extends AbstractFoDJsonNodeOutputCommand implements IRecordTransformer {
     @Getter @Mixin private OutputHelperMixins.Get outputHelper;
-    @Mixin private FoDAppRelResolverMixin.PositionalParameter appRelResolver;
+    @Mixin private FoDReleaseResolverMixin.PositionalParameter appRelResolver;
 
     @Override
     public JsonNode getJsonNode(UnirestInstance unirest) {
